Moved GunSystem interface stuff into GunSystemInterfaces.cs
This clears up GunScript.cs a lot
