+++ UnityProject/Assets/Scripts/GunScriptSystemBlocks.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using ExtentionUtil;
+using GunSystemInterfaces;
 
 namespace GunSystemsV1 {
     [InclusiveAspects(GunAspect.TRIGGER, GunAspect.THUMB_SAFETY)]
