+++ UnityProject/Assets/Scripts/GunSystemInterfaces.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 17a043f43022a384db90957f05ef493a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
